Add sort_images_by option (can be 'date').
My wife and I are both taking photos on our phones. We'd like to be
able to upload the good ones and have them displayed in chronological
order. The filenames are numbered differently, but the EXIF date tags are
pretty accurate, so I'll set sort_images_by => 'date' in our gallery. The
default behavior is unchanged.